summarizer-agent/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
google-adk==0.1.0b4
python-dotenv==1.0.1
pydantic==2.6.3
google-generativeai==0.4.1
requests==2.31.0
